+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 1.0.1
 
+- `WindowMessageLoop.runLoop/runLoopAsync`:
+  - Added `condition`.
 - `pubspec.yaml`: fix repository.
-- `[README.md](README.md)`: usage example.
+- `[README.md](README.md)`:
+  - Usage example.
+  - Added screenshot.
 - Adjust example.
 
 ## 1.0.0